# backend/swagger.json
{
  "info": {
    "title": "Image Uploader",
    "version": "1.0.0-snapshot"
  },
  "basePath": "/api",
  "swagger": "2.0",
  "paths": {
    "/image/{id}": {
      "get": {
        "summary": "Get an image by id",
        "produces": [
          "application/octet-stream"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Image id",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "file"
            }
          },
          "404": {
            "description": "NOT FOUND"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/image": {
      "post": {
        "summary": "Upload an image",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "image",
            "type": "file",
            "description": "file to upload"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "description": "true if successfull"
                },
                "id": {
                  "type": "integer",
                  "description": "id of the uploaded image"
                }
              }
            }
          },
          "422": {
            "description": "Missing file or wrong file size or type",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "default": false,
                  "description": "false to indicate error"
                },
                "message": {
                  "type": "string",
                  "description": "error message"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "definitions": {},
  "responses": {},
  "parameters": {},
  "securityDefinitions": {},
  "tags": []
}